+++ containers/message-parser/app/main.py
@@ -12,11 +12,11 @@ from app.models import ParseMessageInput
 from app.models import ParseMessageResponse
 from app.models import ParsingSchemaModel
 from app.models import PutSchemaResponse
-from app.phdc.phdc import Address
-from app.phdc.phdc import Name
-from app.phdc.phdc import Patient
-from app.phdc.phdc import PHDCBuilder
-from app.phdc.phdc import PHDCInputData
+from app.phdc.models import Address
+from app.phdc.models import Name
+from app.phdc.models import Patient
+from app.phdc.builder import PHDCBuilder
+from app.phdc.models import PHDCInputData
 from app.utils import convert_to_fhir
 from app.utils import DIBBS_REFERENCE_SIGNIFIER
 from app.utils import freeze_parsing_schema
