+++ containers/message-parser/tests/test_phdc.py
@@ -4,12 +4,12 @@ from unittest.mock import patch
 
 import pytest
 from app import utils
-from app.phdc.phdc import Address
-from app.phdc.phdc import Name
-from app.phdc.phdc import Patient
-from app.phdc.phdc import PHDCBuilder
-from app.phdc.phdc import PHDCInputData
-from app.phdc.phdc import Telecom
+from app.phdc.models import Address
+from app.phdc.models import Name
+from app.phdc.models import Patient
+from app.phdc.builder import PHDCBuilder
+from app.phdc.models import PHDCInputData
+from app.phdc.models import Telecom
 from lxml import etree as ET
